Added acctual used repositories to infrastructure layer.
