+++ src/Infrastructure/Doctrine/Repository/DoctrineUserRepository.php
@@ -1,0 +1,22 @@
+<?php
+declare(strict_types=1);
+
+namespace App\Infrastructure\Doctrine\Repository;
+
+use App\DomainModel\Repository\UserRepository;
+use App\Entity\User;
+use Doctrine\Bundle\DoctrineBundle\Repository\ServiceEntityRepository;
+use Doctrine\Persistence\ManagerRegistry;
+
+final class DoctrineUserRepository extends ServiceEntityRepository implements UserRepository
+{
+    public function __construct(ManagerRegistry $registry)
+    {
+        parent::__construct($registry, User::class);
+    }
+
+    public function findOneByApikey(string $apikey): User
+    {
+        return $this->findOneBy(['apikey' => $apikey]);
+    }
+}
